2023:5:2 fix overflow

+++ 2023/src/day5.rs
@@ -2,7 +2,7 @@ use aoc_runner_derive::{aoc, aoc_generator};
 
 use std::str::FromStr;
 
-type Number = u32;
+type Number = u64;
 
 #[derive(Debug, PartialEq)]
 struct Almanac {
